Por convención, hacemos un refactor de implementaciones al paquete _impl_
